Remove OpenOptios::not_cloexec()
Posix message queues are required to be close-on-exec by default,
and passing O_CLOEXEC to mq_open() should have no effect.
Therefore open() doesen't need to consider it, as users
can disable close-on-exec afterwards if they need to.
That is, if NetBSD and DragonFlyBSD wasn't buggy.

This commit also makes is_cloexec() unconditionally available.