Attempt to compile on VxWorks, and disable all non-core features by default
Someone recently added mq functions for VxWorks to libc:
https://github.com/rust-lang/libc/pull/1645
While mqd_t is an int there, that doesn't necessaryily mean that it's a
type of file descriptor: VxWorks is not a unix-like.

libc doesn't have FIOCLEX, FIONCLEX or exec() for VxWorks,
so set_cloexec() had to be disabled there for the crate to compile,
and I assume the concept of close-on-exec isn't meaningful there.

To improve future stability all #[cfg(not(any(x,y,z)))] were changed
into #[cfg(any(a,b,c))]: So that when adding support for another OS,
it won't be a breaking change to disable an impl or method that isn't
correct there but that happened to compile.